Transfer data from the ASIO buffers to the QA401.

+++ src/asio401/ASIO401/asio401.h
@@ -127,7 +127,7 @@ namespace asio401 {
 				const bool host_supports_timeinfo;
 				std::atomic<bool> stopRequested = false;
 				// The index of the "unlocked" buffer (or "half-buffer", i.e. 0 or 1) that contains data not currently being processed by the ASIO host.
-				size_t our_buffer_index;
+				long our_buffer_index;
 				std::atomic<SamplePosition> samplePosition;
 				Win32HighResolutionTimer win32HighResolutionTimer;
